A Loading Screen prefab with adjustable delay for length of animation/ load time.
This comes with some images to use as loading screens as well as all scripts needed for it to work. It works when the rigidbody tagged "Player" ( our player) collides with an object tagged "Door" The animation starts then after animation plays the next scene in the build is called.